Split main.rs: move app logic to app.rs
`main.rs` now contains only the `main` function and module
declarations. All command dispatch helpers (`run`, `run_status`,
`run_compress`, `run_mine`) and utilities (`open_palace`,
`expand_tilde`) move to the new `app` module.

This makes `main.rs` a thin entry point and keeps application
logic in a single, focused module.

By the will of the Machine God
